added optimizations for assembly ect.

+++ src/bc/condition.rs
@@ -13,7 +13,7 @@ pub enum BoundaryConditionType {
 pub trait BoundaryCondition {
     fn apply(&mut self, simulation: &mut Simulation);
     fn initalize(&mut self, simulation: &Simulation){}
-    fn get_nodes(&self) -> &Vec<u32>;
+    fn get_nodes(&self) -> &Vec<usize>;
     fn type_name(&self) -> BoundaryConditionType;
     fn print_stats(&self){}
 }
