+++ src/bc/force.rs
@@ -16,17 +16,17 @@ use crate::bc::BoundaryConditionType;
 /// * `force`: The force to apply to the nodes.
 #[derive(Serialize, Deserialize, Debug)] 
 pub struct LoadCondition {
-    nodes: Vec<u32>,
+    nodes: Vec<usize>,
     force: DVector<f64>
 }
 
 impl LoadCondition {
-    pub fn new(nodes: Vec<u32>, net_force: DVector<f64>) -> Self {
+    pub fn new(nodes: Vec<usize>, net_force: DVector<f64>) -> Self {
         let s = nodes.len() as f64;
         let force = net_force / s;
         LoadCondition { nodes, force }
     }
-    pub fn new_from_vec(nodes: Vec<u32>, net_force: Vec<f64>) -> Self {
+    pub fn new_from_vec(nodes: Vec<usize>, net_force: Vec<f64>) -> Self {
         let force = DVector::from(net_force);
         LoadCondition::new(nodes, force)
     }
@@ -37,13 +37,13 @@ impl BoundaryCondition for LoadCondition {
     fn apply(&mut self, simulation: &mut Simulation) {
         for &node_id in &self.nodes {
             for i in 0..self.force.len() {
-                let global_index = simulation.get_global_index(node_id, i as u32);
-                simulation.load_vector[global_index as usize] += self.force[i];
+                let global_index = simulation.get_global_index(node_id, i );
+                simulation.load_vector[global_index ] += self.force[i];
             }
         }
     }
 
-    fn get_nodes(&self) -> &Vec<u32> {
+    fn get_nodes(&self) -> &Vec<usize> {
         &self.nodes
     }
